+++ Polyhedron/demo/Polyhedron/Plugins/Mesh_3/Mesh_3_plugin_cgal_code.cpp
@@ -44,6 +44,7 @@ Meshing_thread* cgal_code_mesh_3(QList<const SMesh*> pMeshes,
                                  const double tet_sizing,
                                  const double tet_min_sizing,
                                  const double edge_size,
+                                 const double edge_min_size,
                                  const double tet_shape,
                                  bool protect_features,
                                  bool protect_borders,
@@ -56,6 +57,7 @@ Meshing_thread* cgal_code_mesh_3(QList<const SMesh*> pMeshes,
   std::cerr << "Meshing file \"" << qPrintable(filename) << "\"\n";
   std::cerr << "  angle: " << facet_angle << std::endl
             << "  edge size bound: " << edge_size << std::endl
+            << "  edge min size bound: " << edge_min_size << std::endl
             << "  facets size bound: " << facet_sizing << std::endl
             << "  approximation bound: " << facet_approx << std::endl;
   if (!surface_only)
@@ -105,10 +107,12 @@ Meshing_thread* cgal_code_mesh_3(QList<const SMesh*> pMeshes,
             "<ul>"
             "<li>Angle: %1</li>"
             "<li>Edge size bound: %2</li>"
-            "<li>Facets size bound: %3</li>"
-            "<li>Approximation bound: %4</li>")
+            "<li>Edge min size bound: %3</li>"
+            "<li>Facets size bound: %4</li>"
+            "<li>Approximation bound: %5</li>")
     .arg(facet_angle)
     .arg(edge_size)
+    .arg(edge_min_size)
     .arg(facet_sizing)
     .arg(facet_approx);
   if (!surface_only)
@@ -126,6 +130,7 @@ Meshing_thread* cgal_code_mesh_3(QList<const SMesh*> pMeshes,
   param.tet_min_sizing = tet_min_sizing;
   param.tet_shape = tet_shape;
   param.edge_sizing =  edge_size;
+  param.edge_min_sizing = edge_min_size;
   param.manifold = manifold;
   param.protect_features = protect_features || protect_borders;
   param.use_sizing_field_with_aabb_tree = polylines.empty() && protect_features;
@@ -147,6 +152,7 @@ Meshing_thread* cgal_code_mesh_3(const QList<const SMesh*> pMeshes,
                                  const double tet_sizing,
                                  const double tet_min_sizing,
                                  const double edge_size,
+                                 const double edge_min_size,
                                  const double tet_shape,
                                  bool protect_features,
                                  bool protect_borders,
@@ -159,6 +165,7 @@ Meshing_thread* cgal_code_mesh_3(const QList<const SMesh*> pMeshes,
   std::cerr << "Meshing file \"" << qPrintable(filename) << "\"\n";
   std::cerr << "  angle: " << facet_angle << std::endl
     << "  edge size bound: " << edge_size << std::endl
+    << "  edge min size bound: " << edge_min_size << std::endl
     << "  facets size bound: " << facet_sizing << std::endl
     << "  approximation bound: " << facet_approx << std::endl;
   if (!surface_only)
@@ -203,10 +210,12 @@ Meshing_thread* cgal_code_mesh_3(const QList<const SMesh*> pMeshes,
       "<ul>"
       "<li>Angle: %1</li>"
       "<li>Edge size bound: %2</li>"
-      "<li>Facets size bound: %3</li>"
-      "<li>Approximation bound: %4</li>")
+      "<li>Edge min size bound :%3</li>"
+      "<li>Facets size bound: %4</li>"
+      "<li>Approximation bound: %5</li>")
     .arg(facet_angle)
     .arg(edge_size)
+    .arg(edge_min_size)
     .arg(facet_sizing)
     .arg(facet_approx);
   if (!surface_only)
@@ -224,6 +233,7 @@ Meshing_thread* cgal_code_mesh_3(const QList<const SMesh*> pMeshes,
   param.tet_min_sizing = tet_min_sizing;
   param.tet_shape = tet_shape;
   param.edge_sizing = edge_size;
+  param.edge_min_sizing = edge_min_size;
   param.manifold = manifold;
   param.protect_features = protect_features || protect_borders;
   param.use_sizing_field_with_aabb_tree = protect_features;
@@ -246,6 +256,7 @@ Meshing_thread* cgal_code_mesh_3(const Implicit_function_interface* pfunction,
                                  const double tet_sizing,
                                  const double tet_min_sizing,
                                  const double edge_size,
+                                 const double edge_min_size,
                                  const double tet_shape,
                                  const int manifold,
                                  const bool surface_only)
@@ -277,6 +288,7 @@ Meshing_thread* cgal_code_mesh_3(const Implicit_function_interface* pfunction,
   param.tet_min_sizing = tet_min_sizing;
   param.tet_shape = tet_shape;
   param.edge_sizing = edge_size;
+  param.edge_min_sizing = edge_min_size;
   param.manifold = manifold;
   param.detect_connected_components = false; // to avoid random values
                                              // in the debug displays
@@ -306,6 +318,7 @@ Meshing_thread* cgal_code_mesh_3(const Image* pImage,
                                  const double tet_sizing,
                                  const double tet_min_sizing,
                                  const double edge_size,
+                                 const double edge_min_size,
                                  const double tet_shape,
                                  bool protect_features, //detect_polylines
                                  const bool protect_borders,//polylines on bbox
@@ -333,6 +346,7 @@ Meshing_thread* cgal_code_mesh_3(const Image* pImage,
   param.facet_min_sizing = facet_min_sizing;
   param.tet_min_sizing = tet_min_sizing;
   param.edge_sizing = edge_size;
+  param.edge_min_sizing = edge_min_size;
   param.tet_shape = tet_shape;
   param.manifold = manifold;
   param.image_3_ptr = pImage;
@@ -343,10 +357,12 @@ Meshing_thread* cgal_code_mesh_3(const Image* pImage,
       "<ul>"
       "<li>Angle: %1</li>"
       "<li>Edge size bound: %2</li>"
-      "<li>Facets size bound: %3</li>"
-      "<li>Approximation bound: %4</li>")
+      "<li>Edge min size bound: %3</li>"
+      "<li>Facets size bound: %4</li>"
+      "<li>Approximation bound: %5</li>")
     .arg(facet_angle)
     .arg(edge_size)
+    .arg(edge_min_size)
     .arg(facet_sizing)
     .arg(facet_approx);
   if (!surface_only)